Changes to API endpoints
Changes to API endpoints - SV, Generic Assay, Info
